requirements.txt
Flask==3.0.2
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.2.0
torch==2.2.2
torchvision==0.17.2
pytest==8.1.1
